:truck: Move ./test to ./src
